requirements.txt
annotated-types==0.7.0
anthropic==0.68.1
anyio==3.7.1
beautifulsoup4==4.14.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
jiter==0.11.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.0
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
wikipedia==1.4.0
Wikipedia-API==0.6.0